+++ test/clojure_lsp/parser_test.clj
@@ -5,6 +5,18 @@
     [rewrite-clj.zip :as z]
     [taoensso.timbre :as log]))
 
+(deftest find-last-by-pos
+  (is (= "foo/bar" (z/string (parser/find-last-by-pos (z/of-string "(ns foo) (foo/bar 1)")
+                                                      {:row 1 :col 11 :end-row 1 :end-col 11}))))
+  (is (= "1" (z/string (parser/find-last-by-pos (z/of-string "(ns foo) 1 #(+ 1 2) 3")
+                                                {:row 1 :col 10 :end-row 1 :end-col 10}))))
+  (is (= "3" (z/string (parser/find-last-by-pos (z/of-string "(ns foo) 1 #(+ 1 2) 3")
+                                                {:row 1 :col 21 :end-row 1 :end-col 21}))))
+  (is (= "1" (z/string (parser/find-last-by-pos (z/of-string "(ns foo) 1 #?(+ 1 2) 3")
+                                                {:row 1 :col 10 :end-row 1 :end-col 10}))))
+  (is (= "3" (z/string (parser/find-last-by-pos (z/of-string "(ns foo) 1 #?(+ 1 2) 3")
+                                                {:row 1 :col 22 :end-row 1 :end-col 22})))))
+
 (deftest safe-zloc-of-string
   (is (= "(ns foo) foo/bar" (z/string (z/up (#'parser/safe-zloc-of-string "(ns foo) foo/bar")))))
   (is (= "(ns foo) foo/ (+ 1 2)" (z/string (#'parser/safe-zloc-of-string "(ns foo) foo/ (+ 1 2)"))))