+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Implement support for textDocument/signatureHelp - Fixes #324
 - Disable notify-references on didChange with a flag `notify-references-on-change`.
+- Fix completion not working when reader macro on file - Fixes #332
 
 ## 2021.02.19-23.08.40